# Conflicts:
#	src/main/java/gr/etherasTickets/user/CustomUserRepository.java

+++ src/main/java/gr/etherasTickets/user/UserController.java
@@ -1,21 +1,13 @@
 package gr.etherasTickets.user;
 
-import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
-
-import javax.servlet.http.HttpServletResponse;
 
 @RestController
 @RequestMapping("/users")
@@ -38,26 +30,29 @@ public class UserController {
         jsonResponse.put("_userId", response);
         return new ResponseEntity<>(jsonResponse.toString(), HttpStatus.OK);
     }
-    
 
-    @RequestMapping(value = "/new" , method = RequestMethod.POST)
-    public ResponseEntity<String> createUser(@RequestParam Map<String,String> requestParams) throws Exception{
-        ObjectId id = new ObjectId();
-        String firstname=requestParams.get("firstname");
-        String lastname=requestParams.get("lastname");
-        String email=requestParams.get("email");
-        String username=requestParams.get("username");
-        String password=requestParams.get("password");
 
-        repository.createUser(id.toString(),firstname,lastname,email,username,password, 0, new ArrayList<>());
+    @RequestMapping(method = RequestMethod.POST)
+    public ResponseEntity<String> createUser(@RequestBody User user) throws Exception{
+        User users = repository.findByUsername(user.getFirstName());
+        if(users == null){
+            throw new Exception("User exists");
+        }
+        User newUser = user;
+        repository.createUser(newUser);
 
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
-    
-    @RequestMapping(value = "/remove" , method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<String> removeUser(@RequestParam(name="id" , required=true) String id){
-    	repository.removeUser(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+
+    @RequestMapping(path = "/{id}" , method = RequestMethod.DELETE)
+    public ResponseEntity<User> removeUser(@PathVariable String id){
+        User user = repository.getUserById(id);
+        if(user == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        repository.removeUser(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
